スクロール・ハイライト調整ロジックをuseOnboardingに集約
コンポーネントとフックの責務を明確化するため、
ステップ遷移時のスクロール位置調整とハイライト再計算の
useEffectロジックをOnboardingCardからuseOnboardingフックに移動した。
